requirements.txt
langchain 
langgraph 
langchain-community 
langchain-core 
langchain-openai 
langchain-groq 
langchain-exa 
python-dotenv 
streamlit